some changes for models and migrations

+++ database/migrations/2020_02_04_094041_create_portfolio_images_table.php
@@ -15,7 +15,7 @@ class CreatePortfolioImagesTable extends Migration
     {
         Schema::create('portfolio_images', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('portfolio_id');
+            $table->unsignedBigInteger('portfolio_id');
             $table->string('file');
             $table->string('title')->nullable();
             $table->string('description')->nullable();
